requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
dataclasses==0.6
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
mongoengine==0.29.1
packaging==24.2
passlib==1.7.4
pillow==11.2.1
PyJWT==2.9.0
pymongo==3.12.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
